refactor: split monitor watcher
